requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-jose
passlib[bcrypt]
python-multipart
